Change UI to use new ProjectsStore
