+++ src/frontend/ui/base/ProjectBar.svelte
@@ -1,9 +1,17 @@
 <script lang="ts">
-  import { projectsStore } from "@frontend/lib/stores/ProjectBarStore";
+  import { projectsStore } from "@frontend/lib/stores/ProjectStore";
+  // import Shortcuts from "../utils/Shortcuts.svelte";
+  // import { registerShortcuts } from "../../lib/Shortcuts";
 
   let count = 0;
   function createProject() {
     projectsStore.createProject();
+  }
+
+  function draggable(node: HTMLElement) {
+    // let state = params;
+    node.draggable = true;
+    node.style.cursor = "grab";
   }
 
   // projectsStore.subscribe((state) => {
@@ -21,6 +29,16 @@
   function changeName(id: string) {
     projectsStore.changeName(`Hello World ${count++}`, id);
   }
+
+  // TODO: Fix when shortcuts is fixed
+  // registerShortcuts({
+  //   "blix.projects.newProject": () => {
+  //     projectsStore.createProject();
+  //   },
+  //   "blix.projects.closeActiveProject": () => {
+  //     projectsStore.closeProject($projectsStore.activeProject?.id || "");
+  //   },
+  // });
 </script>
 
 <div class="drag flex h-full flex-row flex-nowrap items-center">
@@ -33,6 +51,7 @@
       on:keypress="{() => projectsStore.setActiveProject(project.id)}"
       on:dblclick="{() => changeName(project.id)}"
       on:contextmenu|preventDefault="{() => subscribeToProject(project.id)}"
+      use:draggable
     >
       <p class="mr-2 truncate">{project.name}</p>
       <svg
@@ -61,6 +80,8 @@
     </svg>
   </div>
 </div>
+
+<!-- <Shortcuts {shortcuts} /> -->
 
 <style lang="postcss">
   .no-drag {
